requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
black==22.6.0
click==8.1.3
Deprecated==1.2.13
dj-drf-utils==1.1.2
Django==4.1
django-redis==5.2.0
djangorestframework==3.13.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
pyparsing==3.0.9
pytz==2022.2.1
redis==4.3.4
sqlparse==0.4.2
tomli==2.0.1
wrapt==1.14.1
